% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downread_csv.R
\name{downread_csv}
\alias{downread_csv}
\title{Download and Read a .csv File}
\usage{
downread_csv(path, url)
}
\arguments{
\item{path}{A character string with the name where the downloaded file is saved.}

\item{url}{A character string naming the URL of a resource to be downloaded.}
}
\value{
A tibble
}
\description{
This function will check to see whether the csv file has been downloaded, and
if not it will download the file. It will then load the file using the read_csv
function.
}
\examples{
downread_csv('~/iris.csv', 'https://raw.githubusercontent.com/vincentarelbundock/Rdatasets/master/csv/datasets/iris.csv')

}

