% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper_first.R
\name{toupper_first}
\alias{toupper_first}
\title{Uppercase the First Character of Each Word in a String}
\usage{
toupper_first(string)
}
\arguments{
\item{string}{String to modify}
}
\value{
Returns a string with the first character of each word capitalized, if that character is alphabetic.
}
\description{
This function takes in a string of arbitrary length and returns a string of the 
same length with the first character in each word capitalized, assuming that 
character is alphabetic
}
\examples{
toupper_first('new york')
## -> [1] "New York"

toupper_first('NEW YORK')
## -> [1] "New York"

toupper_first('5th avenue')
## -> [1] "5th Avenue"

}

